requirements.txt
appdirs==1.4.4
asgiref==3.3.0
certifi==2020.11.8
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.3
django-environ==0.4.5
django-ipware==3.0.2
djangorestframework==3.12.2
djangorestframework-gis==0.16
filelock==3.0.12
importlib-metadata==2.0.0
pipenv==2020.11.4
psycopg2==2.8.6
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
virtualenv==20.1.0
virtualenv-clone==0.5.4
whitenoise==5.2.0
zipp==3.4.0
